requirements.txt
# requirements.txt for local LLM with Ollama

# Core dependencies
chromadb>=0.4.22
langchain>=0.1.0
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
requests>=2.31.0

# For PDF processing
pypdf>=4.0.1

# Optional: for enhanced vector search
sentence-transformers>=2.2.2




# Web application
flask

# Database
sqlite3

